requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.3.4
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
fake==0.8
Faker==21.0.0
idna==3.6
iniconfig==2.0.0
invoke==2.2.0
packaging==23.2
paramiko==3.4.0
pip==23.3.2
pluggy==1.3.0
pycparser==2.21
PyNaCl==1.5.0
pytest-cov==4.1.0
pytest-lazy-fixture==0.6.3
pytest==7.4.3
python-dateutil==2.8.2
requests==2.31.0
setuptools==60.2.0
six==1.16.0
urllib3==2.1.0
wheel==0.37.1
wrapt==1.16.0

selenium~=4.16.0
